Sorting Algorithm Visualizer/requirement.txt
absl-py==2.2.2
accelerate==1.10.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.4
altgraph==0.17.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
attrs==25.1.0
audioread==3.0.1
av==16.0.1
Brotli==1.1.0
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
Cython==3.0.12
datasets==4.0.0
decorator==5.2.1
demucs==4.0.1
diffq==0.2.4
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
docstring_parser==0.17.0
dora_search==0.1.12
dpkt==1.9.8
einops==0.8.0
et_xmlfile==2.0.0
fastapi==0.115.12
ffmpeg-python==0.2.0
ffmpy==0.6.3
filelock==3.18.0
fire==0.7.1
flake8==7.1.1
fonttools==4.58.0
frozenlist==1.7.0
fsspec==2025.3.0
future==1.0.0
gradio==5.45.0
gradio_client==1.13.0
graphviz==0.20.3
greenlet==3.2.4
groovy==0.1.2
h11==0.14.0
h5py==3.13.0
hf_transfer==0.1.9
holidays==0.77
holidayskr==0.2.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.35.3
hydra-colorlog==1.2.0
hydra-core==1.3.2
idna==3.10
imbalanced-learn==0.13.0
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.6
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
julius==0.2.7
keras==3.9.2
kiwisolver==1.4.8
korean-lunar-calendar==0.3.1
lameenc==1.8.1
lazy_loader==0.4
librosa==0.11.0
lightgbm==4.6.0
lightning==2.5.2
lightning-utilities==0.15.0
-e git+https://github.com/hiyouga/LLaMA-Factory.git@c867e280937b71904ce20e5b99d562bf9faf042b#egg=llamafactory
llvmlite==0.44.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mccabe==0.7.0
mdurl==0.1.2
ml_dtypes==0.5.1
modelscope==1.30.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.6.3
multiprocess==0.70.16
musdb==0.4.2
museval==0.4.1
mypy==1.15.0
mypy-extensions==1.0.0
mysqlclient==2.2.4
namex==0.0.8
networkx==3.4.2
nltk==3.9.2
numba==0.61.2
numpy==1.26.4
omegaconf==2.3.0
openai==1.30.1
openai-whisper==20240930
opencv-python==4.10.0.84
openpyxl==3.1.5
openunmix==1.3.0
optree==0.15.0
optuna==4.4.0
optuna-integration==4.4.0
orjson==3.10.0
packaging==25.0
pandas==2.3.1
pefile==2023.2.7
peft==0.17.1
pillow==11.2.1
platformdirs==4.5.0
pooch==1.8.2
propcache==0.3.2
protobuf==6.32.1
psutil==7.1.0
pyaml==25.1.0
pyarrow==21.0.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.27.2
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.19.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytorch-forecasting==1.4.0
pytorch-lightning==2.5.2
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rich==14.0.0
rouge-chinese==1.0.3
rpds-py==0.22.3
ruff==0.14.0
safehttpx==0.1.6
safetensors==0.5.3
scapy==2.6.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.1
shellingham==1.5.4
shtab==1.7.2
simplejson==3.19.3
six==1.17.0
sklearn-compat==0.1.3
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
SQLAlchemy==2.0.43
sse-starlette==3.0.2
starlette==0.46.2
stempeg==0.2.3
submitit==1.5.2
sympy==1.14.0
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.7.0
torchaudio==2.1.2
torchmetrics==1.8.0
torchvision==0.22.0
tqdm==4.66.4
transformers==4.57.0
treetable==0.2.5
trl==0.9.6
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tyro==0.8.14
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.2
watchfiles==0.21.0
websockets==12.0
xgboost==3.0.0
xxhash==3.6.0
yarl==1.20.1
